requirements.txt
streamlit>=1.28.0
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
duckdb>=0.9.0
rapidfuzz>=3.5.0
python-dotenv>=1.0.0
openai>=1.12.0
pydeck>=0.8.0
pandas>=2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
requests>=2.31.0
openpyxl>=3.1.0
tqdm>=4.66.0
watchdog>=3.0.0

# Document processing
python-docx>=1.1.0
PyPDF2>=3.0.0

# NLP processing (optional but recommended for QC analysis)
spacy>=3.7.0

# Error Tracking & Monitoring
sentry-sdk>=1.40.0

# Enhanced Logging (optional but recommended)
python-json-logger>=2.0.7

